+++ SOUP.user.js
@@ -3,7 +3,7 @@
 // @namespace   https://github.com/vyznev/
 // @description Miscellaneous client-side fixes for bugs on Stack Exchange sites (development)
 // @author      Ilmari Karonen
-// @version     1.29.4
+// @version     1.29.5
 // @copyright   2014, Ilmari Karonen (http://stackapps.com/users/10283/ilmari-karonen)
 // @license     ISC; http://opensource.org/licenses/ISC
 // @match       *://*.stackexchange.com/*
@@ -385,6 +385,18 @@ fixes.mso286009 = {
 	url:	"http://meta.stackoverflow.com/q/286009",
 	sites:	/^(meta\.)?stackoverflow\./,
 	css:	".nav.askquestion { margin-left: 26px }"
+};
+fixes.mso283939 = {
+	title:	"Flag for diamond moderator textarea breaks out of the dialog",
+	url:	"http://meta.stackoverflow.com/q/283939",
+	// should be safe to apply on all sites; sites other than SO already have this exact CSS
+	css:	".mod-attention-subform textarea { font-size: 12px; width: 575px; padding: 5px }"
+};
+fixes.mso283617 = {
+	title:	"The re-design has borked the user flag history page",
+	url:	"http://meta.stackoverflow.com/q/283617",
+	// "body" added to override conflicting SO style; should be safe (and do nothing) on other SE sites
+	css:	"body .bounty-indicator-tab { display: inline-block }"
 };